HR Chatbot/requirements.txt
Flask==3.0.3
flask-cors==4.0.0
python-dotenv==1.0.1
langchain==0.2.1
langchain-openai==0.1.8
langchain-google-genai==1.0.1
langchain-pinecone==0.1.0
pinecone-client==3.2.2
transformers==4.41.1
sentence-transformers==2.7.0
torch==2.3.0
PyPDF2==3.0.1
huggingface-hub==0.22.2

